refactor: remove product flavors, optimize database rankings, and enhance update installation
- Consolidated the build system by removing `playstore` and `fdroid` flavors, moving shared logic and `PaxsenixSearchHelper` to the main source set.
- Implemented automatic app update installation using `UpdateDownloadReceiver`, `FileProvider`, and the `REQUEST_INSTALL_PACKAGES` permission.
- Optimized `RankingDao` performance by replacing expensive subqueries with Common Table Expressions (CTEs) and `ROW_NUMBER()` window functions for statistics calculation.
- Refactored coroutine handling to use structured concurrency, replacing `GlobalScope` with `lifecycleScope`, `rememberCoroutineScope`, or managed `CoroutineScope` instances.
- Enhanced `PlaybackService` stability with better property initialization checks, improved metadata injection for car displays, and robust error handling for scrobbling services.
- Added support for 24-bit PCM audio processing in `BalanceAudioProcessor` and `ReplayGainAudioProcessor`.
- Optimized `ImprovedShuffleOrder` and `AdvancedForwardingPlayer` using efficient index mapping and caching for upcoming items.
- Standardized network request behavior across lyrics providers (Genius, LRCLib, BetterLyrics, Lyrically) with consistent timeouts and error handling.
- Refactored `MainModule` to use singletons for DAO definitions and introduced a named `applicationScope` for global background tasks.
- Improved database migration logic and added schema version 16 to support additional track metadata.
- Updated `SleepTimer` to support exact alarm scheduling requirements and `CustomArtistImageManager` to implement `AutoCloseable`.
- Adjusted Coil configuration to allow hardware acceleration and updated `AboutScreen` to use auto-mirrored icons.

+++ app/src/main/java/com/mardous/projectmusic/playback/processor/ReplayGainAudioProcessor.kt
@@ -87,7 +87,8 @@ class ReplayGainAudioProcessor(
                 C.ENCODING_PCM_24BIT -> {
                     while (inputBuffer.hasRemaining()) {
                         val sample = inputBuffer.getInt24()
-                        val scaled = (sample * delta).toInt()
+                        val scaled = (sample.toDouble() * delta)
+                            .toInt()
                             .coerceIn(ByteUtils.Int24_MIN_VALUE, ByteUtils.Int24_MAX_VALUE)
                         buffer.putInt24(scaled)
                     }
